requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7
python-pptx==0.6.21
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
openai>=1.5.0 